changes again

+++ GUI_SRC_CODE/src/MAPHEUS/mapheus_main.py
@@ -612,14 +612,15 @@ class MAPHEUS_GUI(QMainWindow, Ui_Title):
             self.p_window_data.terminate()
             self.p_window_data.join()
 
-        #terminate the other subprocess
-        serial_backend.p1.terminate()
-        serial_backend.p1.join()
+        #terminate the other subprocess (may never have started if the device never connected)
+        if serial_backend.p1 is not None:
+            serial_backend.p1.terminate()
+            serial_backend.p1.join()
 
         #stop all the threads
         self.worker_socket.stop()
         self.worker_DataUpdate.stop()
-        
+
         #restart the python script
         os.execv(sys.executable, ['python'] + sys.argv)
 
@@ -685,9 +686,10 @@ class TabWindowMAPHEUS(QMainWindow):
             q.join_thread()
 
 
-        #terminate the other subprocess
-        serial_backend.p1.terminate()
-        serial_backend.p1.join()
+        #terminate the other subprocess (may never have started if the device never connected)
+        if serial_backend.p1 is not None:
+            serial_backend.p1.terminate()
+            serial_backend.p1.join()
 
         #stop all the threads
